requirements.txt
asgiref==3.9.2
Django==5.0.6
django-filter==25.1
djangorestframework==3.15.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
